fix: pulse icon init
